+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/e2e/server/server.cjs
@@ -3,8 +3,8 @@ const express = require('express');
 const server = express();
 const path = require('path');
 const mustacheExpress = require('mustache-express');
-const getAppSettings = require('./src/build/AppSettings.cjs');
-const getDecorator = require('./src/build/decorator.cjs');
+const getAppSettings = require('../../mock/AppSettings.cjs');
+const getDecorator = require('./decorator.cjs');
 const compression = require('compression');
 
 require('dotenv').config();
@@ -13,7 +13,7 @@ server.use(express.json());
 server.disable('x-powered-by');
 server.use(compression());
 
-server.set('views', path.resolve(`${__dirname}/dist`));
+server.set('views', path.resolve(`${__dirname}/../../dist`));
 
 server.set('view engine', 'mustache');
 server.engine('html', mustacheExpress());
@@ -43,7 +43,7 @@ const startServer = async (html) => {
     console.log(process.env.PUBLIC_PATH);
     server.get(`${process.env.PUBLIC_PATH}/health/isAlive`, (_req, res) => res.sendStatus(200));
     server.get(`${process.env.PUBLIC_PATH}/health/isReady`, (_req, res) => res.sendStatus(200));
-    server.use(`${process.env.PUBLIC_PATH}/assets`, express.static(path.resolve(__dirname, 'dist/assets')));
+    server.use(`${process.env.PUBLIC_PATH}/assets`, express.static(path.resolve(__dirname, '../../dist/assets')));
 
     server.get(/^\/(?!.*api)(?!.*dist).*$/, (req, res) => {
         res.send(html);