+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/søknad/steps/legeerklæring/LegeerklæringStep.tsx
@@ -1,3 +1,6 @@
+import { Attachment } from '@navikt/sif-common-core-ds/src/types/Attachment';
+import { getUploadedAttachments } from '@navikt/sif-common-core-ds/src/utils/attachmentUtils';
+import { FormikValuesObserver } from '@navikt/sif-common-formik-ds';
 import { getTypedFormComponents } from '@navikt/sif-common-formik-ds/src/components/getTypedFormComponents';
 import PersistStepFormValues from '../../../components/persist-step-form-values/PersistStepFormValues';
 import { useOnValidSubmit } from '../../../hooks/useOnValidSubmit';
@@ -12,9 +15,6 @@ import SøknadStep from '../../SøknadStep';
 import { getSøknadStepConfigForStep } from '../../søknadStepConfig';
 import LegeerklæringForm, { LegeerklæringFormFields, LegeerklæringFormValues } from './LegeerklæringForm';
 import { getLegeerklæringStepInitialValues, getLegeerklæringSøknadsdataFromFormValues } from './legeerklæringStepUtils';
-import { Attachment } from '@navikt/sif-common-core-ds/src/types/Attachment';
-import { getUploadedAttachments } from '../../../utils/attachmentUtils';
-import { FormikValuesObserver } from '@navikt/sif-common-formik-ds';
 
 const { FormikWrapper } = getTypedFormComponents<LegeerklæringFormFields, LegeerklæringFormValues>();