+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/søknad/steps/situasjon/form-parts/ArbeidsforholdUtbetalingsårsak.tsx
@@ -1,7 +1,9 @@
 import { useEffect, useMemo, useRef, useState } from 'react';
+import { getAttachmentURLFrontend, uploadVedlegg } from '@navikt/sif-common';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 import FormBlock from '@navikt/sif-common-core-ds/src/atoms/form-block/FormBlock';
 import FileUploadErrors from '@navikt/sif-common-core-ds/src/components/file-upload-errors/FileUploadErrors';
+import FormikFileUploader from '@navikt/sif-common-core-ds/src/components/formik-file-uploader/FormikFileUploader';
 import PictureScanningGuide from '@navikt/sif-common-core-ds/src/components/picture-scanning-guide/PictureScanningGuide';
 import { Attachment } from '@navikt/sif-common-core-ds/src/types/Attachment';
 import { getTypedFormComponents, ValidationError } from '@navikt/sif-common-formik-ds';
@@ -12,11 +14,8 @@ import {
 } from '@navikt/sif-common-formik-ds/src/validation';
 import { validateAll } from '@navikt/sif-common-formik-ds/src/validation/validationUtils';
 import { useFormikContext } from 'formik';
-import { ApiEndpoint } from '../../../../api/api';
-import FormikFileUploader from '../../../../components/formik-file-uploader/FormikFileUploader';
 import { useAppIntl } from '../../../../i18n';
 import { Arbeidsforhold, Utbetalingsårsak, ÅrsakNyoppstartet } from '../../../../types/ArbeidsforholdTypes';
-import { fixAttachmentURL } from '../../../../utils/attachmentUtils';
 import { relocateToLoginPage } from '../../../../utils/navigationUtils';
 import { validateAttachments, ValidateAttachmentsErrors } from '../../../../utils/validateAttachments';
 import { AppFieldValidationErrors } from '../../../../utils/validations';
@@ -41,7 +40,7 @@ const ArbeidsforholdUtbetalingsårsak = ({ arbeidsforhold, parentFieldName }: Pr
     const arbeidsgivernavn = arbeidsforhold.navn;
 
     const attachments: Attachment[] = useMemo(() => {
-        return arbeidsforhold ? arbeidsforhold.dokumenter.map(fixAttachmentURL) : [];
+        return arbeidsforhold ? arbeidsforhold.dokumenter : [];
     }, [arbeidsforhold]);
 
     const ref = useRef({ attachments });
@@ -140,7 +139,8 @@ const ArbeidsforholdUtbetalingsårsak = ({ arbeidsforhold, parentFieldName }: Pr
                             attachments={attachments}
                             name={getFieldName(ArbeidsforholdFormFields.dokumenter)}
                             buttonLabel={text('step.situasjon.arbeidsforhold.utbetalingsårsak.vedlegg')}
-                            apiEndpoint={ApiEndpoint.vedlegg}
+                            getAttachmentURLFrontend={getAttachmentURLFrontend}
+                            uploadFile={uploadVedlegg}
                             onErrorUploadingAttachments={setFilesThatDidntGetUploaded}
                             onFileInputClick={() => {
                                 setFilesThatDidntGetUploaded([]);