+++ apps/omsorgspengerutbetaling-soknad/src/app/søknad/steps/legeerklæring/LegeerklæringAttachmentList.tsx
@@ -1,6 +1,6 @@
 import { BodyShort } from '@navikt/ds-react';
 import React from 'react';
-import { vedleggService } from '@navikt/sif-common';
+import { deleteVedlegg } from '@navikt/sif-common';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 import AttachmentListWithDeletion from '@navikt/sif-common-core-ds/src/components/attachment-list-with-deletion/AttachmentListWithDeletion';
 import AttachmentList from '@navikt/sif-common-core-ds/src/components/attachment-list/AttachmentList';
@@ -12,7 +12,6 @@ import {
 import { removeElementFromArray } from '@navikt/sif-common-core-ds/src/utils/listUtils';
 import { useFormikContext } from 'formik';
 import { AppText } from '../../../i18n';
-import { fixAttachmentURL } from '../../../utils/attachmentUtils';
 import { LegeerklæringFormFields, LegeerklæringFormValues } from './LegeerklæringForm';
 
 interface Props {
@@ -25,10 +24,7 @@ const LegeerklæringAvtaleAttachmentList: React.FunctionComponent<Props> = ({
     includeDeletionFunctionality,
 }) => {
     const { values, setFieldValue } = useFormikContext<LegeerklæringFormValues>();
-    const avtale: Attachment[] = values.vedlegg
-        .filter(({ file }: Attachment) => fileExtensionIsValid(file.name))
-        .map(fixAttachmentURL);
-
+    const avtale: Attachment[] = values.vedlegg.filter(({ file }: Attachment) => fileExtensionIsValid(file.name));
     if (!containsAnyUploadedAttachments(avtale)) {
         const noAttachmentsText = (
             <BodyShort>
@@ -51,9 +47,7 @@ const LegeerklæringAvtaleAttachmentList: React.FunctionComponent<Props> = ({
                     const updateFieldValue = () => {
                         setFieldValue(LegeerklæringFormFields.vedlegg, removeElementFromArray(attachment, avtale));
                     };
-                    if (attachment.id) {
-                        vedleggService.delete(attachment.id).then(updateFieldValue, updateFieldValue);
-                    }
+                    deleteVedlegg(attachment).then(updateFieldValue, updateFieldValue);
                 }}
             />
         );