+++ packages/sif-common/src/api/types/innsendingType.ts
@@ -1,3 +1,4 @@
 export enum InnsendingType {
     'omsorgspenger_utbetaling_snf' = '/omsorgspenger-utbetaling-snf/innsending',
+    'omsorgspenger_utbetaling_arbeidstaker' = 'omsorgspenger-utbetaling-arbeidstaker/innsending',
 }